requirements.txt
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
colorama==0.4.6
typing-extensions==4.9.0